IME: another attempt to fix stuck Ctrl modifier in Firefox
41aa7e1 worked for the latest git master of Fcitx5 (commit g4eb394), but
introduced another regression of stuck modifiers with the latest release
of Fcitx5 (5.1.11). I'm not sure about the cause of the regression, but
I guess the commit d9ea016 in Fcitx5 is relevant.

This commit reverts 41aa7e1 and fixes the original issue of stuck Ctrl
modifier in Firefox after Ctrl-F by always calling
wlr_input_method_keyboard_grab_v2_set_keyboard() in
input_method_keyboard_grab_forward_key() even when the unpaired key
events are not forwarded to the IM.

So after this commit, the event flow explained in 41aa7e1 is updated
like this (with Fcitx5 git master):

- The compositor detects F key-release
  - Modifiers (Ctrl pressed) are notified to IM via _set_keyboard()
  - Modifiers (Ctrl pressed) are notified also to the app
  - F key-release is directly notified to the app
- IM sends modifiers (Ctrl pressed) back to the compositor via VK
  - Modifiers (Ctrl pressed) are notified to the app again
- The compositor detects Ctrl key-release
  - Ctrl key-release is directly notified to the app
  - Modifiers (empty) are forwarded to IM
- IM sends modifiers (empty) back to the compositor via VK
  - **Modifier on VK is updated (Ctrl pressed->empty)**
  - Modifiers (empty) are notified to the app