Revert "seat: prevent xwayland from using incorrect keymap (#1816)"
This reverts commit ad15c0474db518ece55846ef624fa2d85358d89e.

This is required for Fcitx5 to see the correct modifier state when it
grabs the keyboard. Previously, the active keyboard on the seat
(wlr_seat.keyboard_state.keyboard) can be NULL when a virtual keyboard
from Fcitx5 is destroyed, which makes it impossible for Fcitx5 to see the
modifier state of the physical keyboard when it grabs the keyboard again.

This revert reintroduces the issue of xwayland clients using incorrect
keymap when keycodes are injected like "wtype a". However, this seems to
be a fundamental issue with X11 that clients cannot get the keymap exactly
at the time when keycodes are received, and is a common problem among
other compositors including sway, wayfire and Hyprland. Specifically for
wtype, the issue can be mitigated by delaying destroying the virtual
keyboard like "wtype -k a -s 200".