action: allow SnapToEdge to combine two cardinal directions
This patch allows to snap a window to e.g. up-left by running two actions:
- `<action name="SnapToEdge" direction="left" />`
- `<action name="SnapToEdge" direction="up" />`

Then running `<action name="SnapToEdge" direction="down" />` snaps it to
left again. This behavior is the same as KWin.

This (and snap-to-adjacent-output feature added in e303281) can be
disabled by adding `relative="no"` in `SnapToEdge`. This means the exact
old behavior (only with snap-to-adjacent-output) cannot be restored, but
I don't think there's much demand for it.