osd-classic: add background color of selected window item
This commit adds new theme options:
- osd.window-switcher.style-classic.item.active.border.color
- osd.window-switcher.style-classic.item.active.bg.color

These theme options configures the border/background of selected window
item in the `classic` style window siwtcher. Previously, the OSD text
color (default: black) was used as the border and background was
transparent.

Now the default theme is updated to use #589bda as the border color
and #c7e2fc as the background color, which are the same as the
`thumbnail` style window switcher.